# improve download performance on fast connections